+++ ContentPatcher/release-notes.md
@@ -1,7 +1,9 @@
 [← back to readme](README.md)
 
 # Release notes
-## Upcoming release
+## 1.9.1
+Released 12 June 2019.
+
 * Fixed error loading local XNB files in some cases with Content Patcher 1.9.
 * Fixed mod-provided tokens being asked for values when they're marked non-ready.